move build_table_output to core
