+++ src/azure-cli-core/azure/cli/core/commands/transform.py
@@ -73,3 +73,28 @@ def gen_dict_to_list_transform(key='value'):
         return result
 
     return _dict_to_list_transform
+
+
+def build_table_output(result, projection):
+
+    if not isinstance(result, list):
+        result = [result]
+
+    final_list = []
+
+    for item in result:
+        def _value_from_path(each_item, path):
+            obj = each_item
+            try:
+                for part in path.split('.'):
+                    obj = obj.get(part, None)
+            except AttributeError:
+                obj = None
+            return obj or ' '
+
+        item_dict = {}
+        for element in projection:
+            item_dict[element[0]] = _value_from_path(item, element[1])
+        final_list.append(item_dict)
+
+    return final_list
